+++ tests/testthat/test-argon2_hash.r
@@ -21,13 +21,17 @@ test_that("Argon2_i password raw hash works", {
   expect_true(typeof(hash$salt) == "raw")
   expect_equal(raw_as_char(hash$raw_hash), toupper(known_good))
 
-  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32, as_raw = F)
+  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32,
+                      as_raw = F)
   expect_true(class(hash) == "argon2.raw")
   expect_true(class(hash$raw_hash) == "argon2.raw.hash")
   expect_true(class(hash$salt) == "argon2.raw.salt")
   expect_true(typeof(hash$raw_hash) == "character")
   expect_true(typeof(hash$salt) == "character")
   expect_equal(as.character(hash$raw_hash), toupper(known_good))
+  expect_identical(hash,
+                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2,
+                               len = 32, as_raw = F))
 
 
   hash <- argon2_hash(pass, 16L, type=type, iterations=5, memory=2, threads=2, len = 32)
@@ -39,7 +43,8 @@ test_that("Argon2_i password raw hash works", {
   expect_true(typeof(hash$salt) == "raw")
   expect_equal(length(hash$salt), 16)
   expect_identical(hash,
-                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2, len = 32))
+                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2,
+                               len = 32))
 
 })
 
@@ -63,13 +68,17 @@ test_that("Argon2_d password raw hash works", {
   expect_true(typeof(hash$salt) == "raw")
   expect_equal(raw_as_char(hash$raw_hash), toupper(known_good))
 
-  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32, as_raw = F)
+  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32,
+                      as_raw = F)
   expect_true(class(hash) == "argon2.raw")
   expect_true(class(hash$raw_hash) == "argon2.raw.hash")
   expect_true(class(hash$salt) == "argon2.raw.salt")
   expect_true(typeof(hash$raw_hash) == "character")
   expect_true(typeof(hash$salt) == "character")
   expect_equal(as.character(hash$raw_hash), toupper(known_good))
+  expect_identical(hash,
+                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2,
+                               len = 32, as_raw = F))
 
   hash <- argon2_hash(pass, 16L, type=type, iterations=5, memory=2, threads=2, len = 32)
   expect_true(class(hash) == "argon2.raw")
@@ -80,7 +89,8 @@ test_that("Argon2_d password raw hash works", {
   expect_true(typeof(hash$salt) == "raw")
   expect_equal(length(hash$salt), 16)
   expect_identical(hash,
-                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2, len = 32))
+                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2,
+                               len = 32))
 })
 
 type <- "id"
@@ -103,13 +113,17 @@ test_that("Argon2_id password raw hash works", {
   expect_true(typeof(hash$salt) == "raw")
   expect_equal(raw_as_char(hash$raw_hash), toupper(known_good))
 
-  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32, as_raw = F)
+  hash <- argon2_hash(pass, nonce, type=type, iterations=5, memory=2, threads=2, len = 32,
+                      as_raw = F)
   expect_true(class(hash) == "argon2.raw")
   expect_true(class(hash$raw_hash) == "argon2.raw.hash")
   expect_true(class(hash$salt) == "argon2.raw.salt")
   expect_true(typeof(hash$raw_hash) == "character")
   expect_true(typeof(hash$salt) == "character")
   expect_equal(as.character(hash$raw_hash), toupper(known_good))
+  expect_identical(hash,
+                   argon2_hash(pass, hash$salt, type=type, iterations=5, memory=2, threads=2,
+                               len = 32, as_raw = F))
 
   hash <- argon2_hash(pass, 16L, type=type, iterations=5, memory=2, threads=2, len = 32)
   expect_true(class(hash) == "argon2.raw")